% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.secretsmanager_package.R
\docType{package}
\name{paws.secretsmanager-package}
\alias{paws.secretsmanager}
\alias{paws.secretsmanager-package}
\title{paws.secretsmanager: AWS Secrets Manager}
\description{
AWS Secrets Manager is a web service that enables
    you to store, manage, and retrieve, secrets.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
